Fix pylint len-as-condition issues
See,
len-as-condition (C1801): Do not use `len(SEQUENCE)` without comparison to determine if a sequence is empty Used when Pylint detects that len(sequence) is being used without explicit comparison inside a condition to determine if a sequence is empty. Instead of coercing the length to a boolean, either rely on the fact that empty sequences are false or compare the length against a scalar.

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>