Fix code issues found by pylint (#160)
* Fix pylint `wrong-import-order` issue

See,
wrong-import-order (C0411): Used when PEP8 import order is not respected (standard imports first, then third-party libraries, then local imports)

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint bad-whitespace issue

bad-whitespace (C0326): Used when a wrong number of spaces is used around an operator, bracket or block opener.

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint trailing-newlines issues

See,
trailing-newlines (C0305):
 	Trailing newlines Used when there are trailing blank lines in a file.

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint bad-continuation issues

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint no-else-return return

See,
no-else-return (R1705): Unnecessary after "return" Used in order to highlight an unnecessary block of code following an if containing a return statement. As such, it will warn when it encounters an else following a chain of ifs, all of them containing a return statement.

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint len-as-condition issues

See,
len-as-condition (C1801): Do not use `len(SEQUENCE)` without comparison to determine if a sequence is empty Used when Pylint detects that len(sequence) is being used without explicit comparison inside a condition to determine if a sequence is empty. Instead of coercing the length to a boolean, either rely on the fact that empty sequences are false or compare the length against a scalar.

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

* Fix pylint missing-final-newline issue

Signed-off-by: Mario Trangoni <mjtrangoni@gmail.com>

Co-authored-by: Mario Trangoni <mjtrangoni@gmail.com>