feat: add rule to check file name of SLS files
Add new rule to check file name of SLS files, as SLS file with a period
in the name (besides the suffix period) can not be referenced.

This commit also allows rules to only run for specific file types by
adding valid `languages` to the rule. File types are determined based
upon the file extension.

Fixes #205

Signed-off-by: Roald Nefs <info@roaldnefs.com>